Begin lib refactor: Model preprocessing
Moved many of the functions used for memo generation in the CDD component into a new lib file called `modelPreprocessing`.
Removed old getIOMapFromModelJSON file. This was a single-function file. Function moved to the new `modelPreprocessing` file.
